Remove pre/post-processing DAMASK envs
The input/output mappers use the same environment in which MatFlow is installed. So as long as a suitable version of the damask Python package is installed in this environment, we don't need to activate the conda environment in the task_preparation/task_processing `env`. In fact, unless we load DAMASK as well, this doesn't work.